Add methods for create and get languages, frameworks and datastorages

+++ src/main/java/com/stack_advisor/stack_advisor_backend/repositories/FrameworkRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface FrameworkRepository extends JpaRepository<Framework, Long> {
+public interface FrameworkRepository extends JpaRepository<Framework, Integer> {
     List<Framework> findByNameIn(List<String> names);
     List<Framework> findByLanguagesContaining(Language language);
 }
